Fix runfiles and sourcemap behavior
This change fixes #116 where certain data files were not being properly
propagated. It also causes raw sources to be embedded in the sourcemap
by default. This is awesome and makes life so much easier, but special
care needs to be taken to not allow the .js.map file to be publicly
accessible.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=140474661