Fix Closure Rules build
These tests were originally designed when Closure Rules existed only on
GitHub and tracked releases of Closure Compiler. When performing
upgrades, it was nice to be able to see when the output of the compiler
changed. However this leads to fragility in Google's shared repository.
So we're going to scale this back.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=145337541