Improve filegroup_external srcs handling
Earlier, a glob would be used if an archive is extracted. Now the srcs
attribute can be set to specify a more restrictive set of files. In
cases where the glob is used, we now ensure the generated BUILD and
WORKSPACE files will not be included in the filegroup.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=155227696